[FLINK-24941][datadog] Move DGauge tests to separate file
